Removed deprecated files + renamed working version
